Move MemoryLayout into the shared compiler vocabulary (#363)
`MemoryLayout` was defined in `core/wasm-codegen`, which is the wrong
crate for it once `pages` and `stack-size` become things a user writes.
`Inference.toml` must be validated when it is loaded, by `apps/infs`,
which cannot reach into code generation to learn what a legal layout is;
the alternative is a second copy of the invariants, and a second copy is
exactly the arrangement the previous commit removed.

`core/compiler-interface` already holds the user-facing vocabulary for
the same reason — `WasmFeatureName` and its diagnostics live there so
that a manifest entry and a command-line flag are rejected in the same
words — and both `apps/infs` and `core/cli` already depend on it. The
type moves there together with `PAGE_SIZE` and `FRAME_ALIGNMENT`, which
`validate` reads: the grid a rejected stack size is measured against
must be the grid code generation rounds frames to, or the error names a
constraint the emitter does not have.

Unlike `WasmFeatureName`, this type is not mirrored on the codegen side.
There, the user's request and code generation's emission permissions are
genuinely different things, and the mapping between them carries a
decision. Here they are the same two numbers, so a mirror would buy
nothing and give the invariants somewhere to drift apart.

`core/wasm-codegen` re-exports it, so `inference_wasm_codegen::MemoryLayout`
and `CodegenOptions.layout` are unchanged and all seventeen references
outside these two crates resolve unedited. No behavior change: same 51
targets, same 2906 tests, every golden byte-identical.
